- Take better advantage of `InputPacket::ang`'s `FRotator` status.
* Change it to a DRotator as that's what the game is expecting. Too many narrowings against DAngle objects needed.
* Utilise object methods where possible.
* Do all math against the object where possible, not its members.

+++ source/core/gamestruct.h
@@ -115,7 +115,7 @@ struct GameInterface
 	virtual void RemoveQAVInterpProps(const int res_id) { }
 	virtual bool WantEscape() { return false; }
 	virtual void StartSoundEngine() = 0;
-	virtual void doPlayerMovement(const float scaleAdjust) = 0;
+	virtual void doPlayerMovement(const double scaleAdjust) = 0;
 	virtual unsigned getCrouchState() = 0;
 
 	virtual FString statFPS()
